`jake btest` was broken.
This happened when accounting.js was added to the thirdparty list.

When `jake btest` is run, Ripple's client is bundled and injected
into the HTML test runner as a giant string. The way the JS string was
being injected was to use String.replace. However, it turns out that
something like `'$'` will trigger a regex feature (unexpectedly..),
and the resulting string that is injected ends right after that
character. In this case, it resulted in a "unexpected end of
input" interpreter error.

By not using regex to inject the JS strings, and instead using
Array.split and Array.join, the strings can be concatenated without
random things like this happening.