The only valid reference material we have is our Project's site.
Eventually, this could point to our Wiki, for one.

Before, it was the BB contributor forum, but since we are purely in ASF
(incubator), this generates some confusion, and points people to the
incorrect place (which is our mailing list).

Also, the doc/ reference link was removed as it is listed in
the website's content.