Remove Images that were not able to licence apache.
When BlackBerry first opensourced Ripple there was some problems
with them open sourcing the images of their phones.  I have removed
all the images that BlackBerry put in of their phones and removed the
notice and Readme.

This should help us get closer to being able to do an Apache release