Fixed a couple of test problems
Updated version of jsdom from 0.6.x to 3.1.0.
The older jsdom fails with a RangeError, saying
"The normalization form should be one of NFC, NFD, NFKC, NFKD."
I tested the newer jsdom with node 0.11.15 and 0.12.0.
I can't get Ripple to install with node 0.10.5, because some modules
require version specifications containing "^", and that form isn't
understood by the version of npm that comes with node 0.10.5.

Also had to update bower to version 1.3.12 for similar reasons.
It wasn't successfully installing jasmine into thirdparty folder.
Finally, fixed a unit test that only passed if two calls to Date retrned
the same value.  The calls are close together so this usually works,
but it does cause intermittant failures.