Fixed code that throws an exception from UI layer
The JavaScript part of the Battery Status panel threw an exception
if the program under test has no <script> tag for cordova.js.
You is not the right way to handle errors in the UI layer.
This will result in an uncaught exception in the emulator, not in the
program under test.  Use an alert instead to communicate to the user.