RIPPLE-91 Fix gpx file implementation in Geolocation panel.
Clicking on the gpx file input box raises a file picker dialog. Selecting a file usually leads to a change event on the input box. This causes loadGpxFile to be called, passing in two null values and an array containing one element, which is the change event. loadGpxFile then extracts the change event from the array, extracts the array of files selected by the file picker from the change event, and reads the file that is the first element of this array.

There are a number of problems with this implementation.
1) if you select a file and then click on the input box again and cancel out of the file picker, then you select no files but still generate a change event.  This causes loadGpxFile to throw an unhandled exception.
2) if an error were to occur while reading the file, loadGpxFile would call a null function value (the "fail" callback).
3) loadGpx has an absurd interface.  This routine is only called in one place.  All you need is a string which is the file to load.

The fix is to verify that a file was selected before calling loadGpxFile, and to have loadGpxFile call console.log to report read errors.

In addition to fixing these problems, I also cleaned up the mixture of CR and CRLF line endings.  This was the only change to geoView/panel.html.

This closes #63