RIPPLE-99 Updated emulation for Vibration API
The Vibration API has added two new native functions: vibrateWithPattern
and cancelVibration.  The vibrateWithPattern function takes an array of
integers and alternately waits and then shakes by the indicated number of
milliseconds.  The cancelVibration function stops the current vibration
action, either initiated by the old vibrate function or
vibrateWithPattern.

Formally vibrateWithPattern requires the shaking to occur for a specific
period of time, rather than just "N shakes".  It's not critical that we
get that exactly right, but it is necessary that the shaking not go on
too long as that can cause two shaking patterns to run into one another.
Accordingly I changed the estimate that one shake will last 160 ms
rather than 100 ms.

This closes #72