Fixed RIPPLE-100 (battery status event emulation)
Fixed the logic so that you get a batterycritical event when the battery
status changes from above 5 to 5 or below, not just when the battery
status changes to exactly 5.  Similarly for batterylow event.  Note that
you must test for batterycritical before checking for batterylow, e.g.
slider move from 25 to 2 causes batterycritical, not batterylow.
Seemed like overkill to trigger both events at the same time.